Change usage of abort to  StarknetDevnetException
