+++ starknet_devnet/blueprints/postman.py
@@ -4,8 +4,9 @@ Postman routes.
 
 import json
 
-from flask import Blueprint, jsonify, request, abort
+from flask import Blueprint, jsonify, request
 from starknet_devnet.state import state
+from starknet_devnet.util import StarknetDevnetException
 
 postman = Blueprint("postman", __name__, url_prefix="/postman")
 
@@ -15,7 +16,8 @@ def validate_load_messaging_contract(request_dict: dict):
     network_url = request_dict.get("networkUrl")
     if network_url is None:
         error_message = "L1 network or StarknetMessaging contract address not specified"
-        abort(400, error_message)
+        raise StarknetDevnetException(message=error_message, status_code=400)
+
     return network_url
 
 @postman.route("/load_l1_messaging_contract", methods=["POST"])
